+++ CHANGELOG.md
@@ -5,18 +5,25 @@ Release Notes
 
 ### New
 
-- `TransactionObserver.ignoreDatabaseChangesUntilNextTransaction` allows transaction observers to stop observing the database for the remaining extent of a transaction.
+- `TransactionObserver.stopObservingDatabaseChangesUntilNextTransaction` allows transaction observers to stop observing the database for the remaining extent of a transaction.
 - GRDB no longer prevents the [truncate optimization](https://www.sqlite.org/lang_delete.html#truncateopt) when no transaction observers are interested in deleted rows.
 - FetchedRecordsController now avoids checking for changes in untracked rowIds.
+- `DatabaseSelectionInfo` (TODO)
 - `Row` adopts RandomAccessCollection
 
+### Deprecated
+
+- `SelectStatement.SelectionInfo` is deprecated, renamed `DatabaseSelectionInfo`.
 
 ### API diff
 
 ```diff
 +extension TransactionObserver {
-+    func ignoreDatabaseChangesUntilNextTransaction()
++    func stopObservingDatabaseChangesUntilNextTransaction()
 +}
+TODO: +DatabaseSelectionInfo
+TODO: +SelectStatement.SelectionInfo
+TODO: -DatabaseEvent.impacts
 ```